+++ ExtendedClasses/ExtendedGroupCounts.cs
@@ -21,6 +21,8 @@ public class GroupCountEntry
 {
     public List<int> counts = [];
     public int weight = 1;
+    public int minPlayerCount = 1;
+    public int maxPlayerCount = 100;
 
     public GroupCountEntry(int i)
     {