+++ Prod/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.2.6
+- Added minimum and maximum player count settings to the possible group counts in `GroupsPerLevel.json`. With this you can make different configs for different player counts. If there are less players than the minimum or more than the maximum set for an entry then that entry can not be selected by the game when loading a level
+- Updated Thunderstore description with info about the new features from 1.2.5 and this version
+
 ### 1.2.5
 - Added new spawn group weight multipliers for each level type (museum, manor, arctic and wizard). The property looks like this: `"levelWeightMultipliers": {"Arctic": 1.0, "Manor": 1.0, "Museum": 1.0, "Wizard": 1.0}`. All three weights of a spawn group will be multiplied by the number corresponding to the current level type. Example use cases: If you set "Manor" to 0 for a group then the group will not be able to spawn in the headman manor. If you set it to 0.5 for museum then the group will be half as likely to spawn on museum compared to the other levels. I decided to do it this way instead of a simple black- or whitelist to allow for more fine-tuning
 - New / modded levels will hopefully be added to this list automatically but we'll see once they work again